Move 'inline' actions to RefactorProcessor
